Refactor common subprocess interactions
